Revert "azure: Initial windows node support"
Automating WinRM for performing the agent bootstrap is unlikely to be
realistic due to the fairly significant security prerequisites described
at https://learn.microsoft.com/en-us/azure/virtual-machines/windows/connect-winrm.
OpenSSH is easier to enable automatically on Windows, but is not
compatible with the Puppet Orchestrator's SSH transport. More research
is required.

This reverts commit 21eb9777cf3b4d66374317b71eccd0abd3851596.